Add checks to controllers.this improved debugging and increased the responsiveness of the rest api

+++ src/main/java/com/example/phonebook/controller/PhoneBookController.java
@@ -16,11 +16,17 @@ public class PhoneBookController {
 
     @GetMapping("/entries/{phone}")
     public ResponseEntity<Entry> getEntryByPhone(@PathVariable("phone") String phone) {
+        if (phone == null) {
+            return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+        }
         return new ResponseEntity<>(userList.getEntryByPhone(phone), HttpStatus.OK);
     }
 
     @PostMapping("/entries/users/{id}")
     public ResponseEntity<List<Entry>> updateEntry(@PathVariable("id") String id, @RequestBody Entry entry) {
+        if (id == null || entry == null) {
+            return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+        }
         userList.addEntry(id, entry);
         return new ResponseEntity<>(userList.getUser(id).getPhoneBookList(), HttpStatus.CREATED);
     }
@@ -28,12 +34,22 @@ public class PhoneBookController {
     @GetMapping("/users/{userId}/entries/{entryId}")
     public ResponseEntity<Entry> getEntry(@PathVariable("userId") String userId,
                                           @PathVariable("entryId") String entryId) {
-        return new ResponseEntity<>(userList.getEntryById(userId, entryId), HttpStatus.OK);
+        if (userId == null || entryId == null) {
+            return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+        }
+        Entry entry = userList.getEntryById(userId, entryId);
+        if (entry == null) {
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
+        return new ResponseEntity<>(entry, HttpStatus.OK);
     }
 
     @DeleteMapping("/users/{userId}/entries/{entryId}")
     public ResponseEntity<Entry> deleteEntry(@PathVariable("userId") String userId,
                                              @PathVariable("entryId") String entryId) {
+        if (userId == null || entryId == null) {
+            return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+        }
         userList.deleteEntryById(userId, entryId);
         return new ResponseEntity<>(HttpStatus.NO_CONTENT);
     }
@@ -42,6 +58,9 @@ public class PhoneBookController {
     public ResponseEntity<List<Entry>> updateUser(@PathVariable("userId") String userId,
                                                   @PathVariable("entryId") String entryId,
                                                   @RequestBody Entry entry) {
+        if (userId == null || entryId == null || entry == null) {
+            return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+        }
         userList.updateEntry(userId, entryId, entry);
         return new ResponseEntity<>(userList.getAllEntryOfUser(userId), HttpStatus.OK);
     }
